Downloads/subspageeval-main/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
lxml==4.9.3
playwright==1.40.0
flask==3.0.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.26.2
anthropic==0.18.1
python-dotenv==1.0.0
langdetect==1.0.9